+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Maven Plugin Utils Changelog
 
+## 1.3.0
+Features:
+* Add protected method [AntFileScanner#createDirectoryScanner()](http://maven-plugin-utils.projects.gabrys.biz/LATEST/apidocs/index.html?biz/gabrys/maven/plugin/util/io/AntFileScanner.html)
+* Add protected method [RegexFileScanner#createFileFilter(File, String[], String[])](http://maven-plugin-utils.projects.gabrys.biz/LATEST/apidocs/index.html?biz/gabrys/maven/plugin/util/io/RegexFileScanner.html)
+
+Bugs:
+* Fixed incorrect implementation of method [Time#toString()](http://maven-plugin-utils.projects.gabrys.biz/LATEST/apidocs/index.html?biz/gabrys/maven/plugin/util/timer/Time.html)
+
 ## 1.2
 Features:
 * Allow access to all classes from [biz.gabrys.maven.plugin.util.io](http://maven-plugin-utils.projects.gabrys.biz/1.2/apidocs/index.html?biz/gabrys/maven/plugin/util/io/package-summary.html) package